+++ BizHawk.Emulation.Cores/Consoles/Sega/gpgx/GPGX.cs
@@ -2,9 +2,6 @@
 using System.Runtime.InteropServices;
 
 using BizHawk.Emulation.Common;
-using BizHawk.Common;
-using System.IO;
-using BizHawk.Emulation.Common.BizInvoke;
 
 namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 {
@@ -15,13 +12,12 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 		isReleased: true,
 		portedVersion: "r874",
 		portedUrl: "https://code.google.com/p/genplus-gx/",
-		singleInstance: false
+		singleInstance: true
 		)]
 	public partial class GPGX : IEmulator, ISyncSoundProvider, IVideoProvider, ISaveRam, IStatable, IRegionable,
 		IInputPollable, IDebuggable, IDriveLight, ICodeDataLogger, IDisassemblable
-	{
-		LibGPGX Core;
-		ElfRunner Elf;
+    {
+		static GPGX AttachedCore = null;
 
 		DiscSystem.Disc CD;
 		DiscSystem.DiscSectorReader DiscSectorReader;
@@ -54,7 +50,7 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 		public GPGX(CoreComm comm, byte[] rom, DiscSystem.Disc CD, object Settings, object SyncSettings)
 		{
 			ServiceProvider = new BasicServiceProvider(this);
-			// this can influence some things internally (autodetect romtype, etc)
+			// this can influence some things internally
 			string romextension = "GEN";
 
 			// three or six button?
@@ -68,27 +64,22 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 
 			try
 			{
-				Elf = new ElfRunner(Path.Combine(comm.CoreFileProvider.DllPath(), "gpgx.elf"), 8 * 1024 * 1024, 36 * 1024 * 1024, 4 * 1024 * 1024);
-				if (Elf.ShouldMonitor)
-					Core = BizInvoker.GetInvoker<LibGPGX>(Elf, Elf);
-				else
-					Core = BizInvoker.GetInvoker<LibGPGX>(Elf);
-
 				_syncSettings = (GPGXSyncSettings)SyncSettings ?? new GPGXSyncSettings();
 				_settings = (GPGXSettings)Settings ?? new GPGXSettings();
 
 				CoreComm = comm;
+				if (AttachedCore != null)
+				{
+					AttachedCore.Dispose();
+					AttachedCore = null;
+				}
+				AttachedCore = this;
 
 				LoadCallback = new LibGPGX.load_archive_cb(load_archive);
 
 				this.romfile = rom;
 				this.CD = CD;
-				if (CD != null)
-				{
-					this.DiscSectorReader = new DiscSystem.DiscSectorReader(CD);
-					cd_callback_handle = new LibGPGX.cd_read_cb(CDRead);
-					Core.gpgx_set_cdd_callback(cd_callback_handle);
-				}
+				this.DiscSectorReader = new DiscSystem.DiscSectorReader(CD);
 
 				LibGPGX.INPUT_SYSTEM system_a = LibGPGX.INPUT_SYSTEM.SYSTEM_NONE;
 				LibGPGX.INPUT_SYSTEM system_b = LibGPGX.INPUT_SYSTEM.SYSTEM_NONE;
@@ -128,20 +119,28 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 				}
 
 
-				if (!Core.gpgx_init(romextension, LoadCallback, _syncSettings.UseSixButton, system_a, system_b, _syncSettings.Region, _settings.GetNativeSettings()))
+				if (!LibGPGX.gpgx_init(romextension, LoadCallback, _syncSettings.UseSixButton, system_a, system_b, _syncSettings.Region, _settings.GetNativeSettings()))
 					throw new Exception("gpgx_init() failed");
 
 				{
 					int fpsnum = 60;
 					int fpsden = 1;
-					Core.gpgx_get_fps(ref fpsnum, ref fpsden);
+					LibGPGX.gpgx_get_fps(ref fpsnum, ref fpsden);
 					CoreComm.VsyncNum = fpsnum;
 					CoreComm.VsyncDen = fpsden;
 					Region = CoreComm.VsyncRate > 55 ? DisplayType.NTSC : DisplayType.PAL;
 				}
 
 				// compute state size
-				InitStateBuffers();
+				{
+					byte[] tmp = new byte[LibGPGX.gpgx_state_max_size()];
+					int size = LibGPGX.gpgx_state_size(tmp, tmp.Length);
+					if (size <= 0)
+						throw new Exception("Couldn't Determine GPGX internal state size!");
+					_savebuff = new byte[size];
+					_savebuff2 = new byte[_savebuff.Length + 13];
+					Console.WriteLine("GPGX Internal State Size: {0}", size);
+				}
 
 				SetControllerDefinition();
 
@@ -151,7 +150,7 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 				SetMemoryDomains();
 
 				InputCallback = new LibGPGX.input_cb(input_callback);
-				Core.gpgx_set_input_callback(InputCallback);
+				LibGPGX.gpgx_set_input_callback(InputCallback);
 
 				if (CD != null)
 					DriveLightEnabled = true;
@@ -167,8 +166,6 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 
 				Tracer = new GPGXTraceBuffer(this, MemoryDomains, this);
 				(ServiceProvider as BasicServiceProvider).Register<ITraceable>(Tracer);
-
-				Elf.Seal();
 			}
 			catch
 			{
@@ -311,6 +308,8 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 			LibGPGX.CDData ret = new LibGPGX.CDData();
 			int size = Marshal.SizeOf(ret);
 
+			ret.readcallback = cd_callback_handle = new LibGPGX.cd_read_cb(CDRead);
+
 			var ses = CD.Session1;
 			int ntrack = ses.InformationTrackCount;
 
@@ -354,7 +353,7 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 		private void SetControllerDefinition()
 		{
 			inputsize = Marshal.SizeOf(typeof(LibGPGX.InputData));
-			if (!Core.gpgx_get_control(input, inputsize))
+			if (!LibGPGX.gpgx_get_control(input, inputsize))
 				throw new Exception("gpgx_get_control() failed");
 
 			ControlConverter = new GPGXControlConverter(input);
@@ -370,8 +369,8 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 
 		public void UpdateVDPViewContext(LibGPGX.VDPView view)
 		{
-			Core.gpgx_get_vdp_view(view);
-			Core.gpgx_flush_vram(); // fully regenerate internal caches as needed
+			LibGPGX.gpgx_get_vdp_view(view);
+			LibGPGX.gpgx_flush_vram(); // fully regenerate internal caches as needed
 		}
 
 		short[] samples = new short[4096];
@@ -392,7 +391,7 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 		void update_audio()
 		{
 			IntPtr src = IntPtr.Zero;
-			Core.gpgx_get_audio(ref nsamp, ref src);
+			LibGPGX.gpgx_get_audio(ref nsamp, ref src);
 			if (src != IntPtr.Zero)
 			{
 				Marshal.Copy(src, samples, 0, nsamp * 2);
