+++ BizHawk.Emulation.Cores/Consoles/Sega/gpgx/GPGXControlConverter.cs
