+++ BizHawk.Emulation.Cores/Consoles/Sega/gpgx/LibGPGX.cs
@@ -3,25 +3,22 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Runtime.InteropServices;
-using BizHawk.Emulation.Common.BizInvoke;
 
 namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 {
-	public abstract class LibGPGX
+	public static class LibGPGX
 	{
-		public const string DllName = "libgenplusgx.dll";
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_get_video(out int w, out int h, out int pitch, ref IntPtr buffer);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_get_video(out int w, out int h, out int pitch, ref IntPtr buffer);
-
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_get_audio(ref int n, ref IntPtr buffer);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_get_audio(ref int n, ref IntPtr buffer);
 
 		[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 		public delegate int load_archive_cb(string filename, IntPtr buffer, int maxsize);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_advance();
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_advance();
 
 		public enum Region : int
 		{
@@ -51,36 +48,50 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 			public uint BackdropColor;
 		}
 
-		[BizImport(CallingConvention.Cdecl, Compatibility=true)]
-		public abstract bool gpgx_init(string feromextension, load_archive_cb feload_archive_cb, bool sixbutton, INPUT_SYSTEM system_a, INPUT_SYSTEM system_b, Region region, [In]InitSettings settings);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern bool gpgx_init(string feromextension, load_archive_cb feload_archive_cb, bool sixbutton, INPUT_SYSTEM system_a, INPUT_SYSTEM system_b, Region region, [In]InitSettings settings);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_get_fps(ref int num, ref int den);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_get_fps(ref int num, ref int den);
 
-		[BizImport(CallingConvention.Cdecl, Compatibility = true)]
-		public abstract bool gpgx_get_control([Out]InputData dest, int bytes);
-		[BizImport(CallingConvention.Cdecl, Compatibility = true)]
-		public abstract bool gpgx_put_control([In]InputData src, int bytes);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern int gpgx_state_max_size();
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern int gpgx_state_size(byte[] dest, int size);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern bool gpgx_state_save(byte[] dest, int size);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern bool gpgx_state_load(byte[] src, int size);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract IntPtr gpgx_get_sram(ref int size);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern bool gpgx_get_control([Out]InputData dest, int bytes);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern bool gpgx_put_control([In]InputData src, int bytes);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract bool gpgx_put_sram(byte[] data, int size);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_get_sram(ref IntPtr area, ref int size);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_clear_sram();
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_clear_sram();
 
 		public const int MIN_MEM_DOMAIN = 0;
 		public const int MAX_MEM_DOMAIN = 13;
 
-		[BizImport(CallingConvention.Cdecl)]
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
 		// apparently, if you use built in string marshalling, the interop will assume that
 		// the unmanaged char pointer was allocated in hglobal and try to free it that way
-		public abstract IntPtr gpgx_get_memdom(int which, ref IntPtr area, ref int size);
+		public static extern IntPtr gpgx_get_memdom(int which, ref IntPtr area, ref int size);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_reset(bool hard);
+		// call this before reading sram returned by gpgx_get_sram()
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_sram_prepread();
+
+		// call this after writing sram returned by gpgx_get_sram()
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_sram_commitwrite();
+
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_reset(bool hard);
 
 		public const int MAX_DEVICES = 8;
 
@@ -138,8 +149,8 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 		[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 		public delegate void input_cb();
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_set_input_callback(input_cb cb);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_set_input_callback(input_cb cb);
 
 		[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 		public delegate void mem_cb(uint addr);
@@ -147,11 +158,11 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 		[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 		public delegate void CDCallback(int addr, CDLog_AddrType addrtype, CDLog_Flags flags);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_set_mem_callback(mem_cb read, mem_cb write, mem_cb exec);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_set_mem_callback(mem_cb read, mem_cb write, mem_cb exec);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_set_cd_callback(CDCallback cd);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_set_cd_callback(CDCallback cd);
 
 
 
@@ -274,10 +285,8 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 			public int last;
 			[MarshalAs(UnmanagedType.ByValArray, SizeConst = CD_MAX_TRACKS)]
 			public readonly CDTrack[] tracks = new CDTrack[CD_MAX_TRACKS];
+			public cd_read_cb readcallback;
 		}
-
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_set_cdd_callback(cd_read_cb cddcb);
 
 		[StructLayout(LayoutKind.Sequential)]
 		public struct VDPNameTable
@@ -298,23 +307,14 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 			public VDPNameTable NTW;
 		}
 
-		[BizImport(CallingConvention.Cdecl, Compatibility = true)]
-		public abstract void gpgx_get_vdp_view([Out] VDPView view);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_get_vdp_view([Out] VDPView view);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_poke_vram(int addr, byte value);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_poke_vram(int addr, byte value);
 
-		/// <summary>
-		/// regenerate whatever portions of the bg pattern cache are currently dirty.
-		/// </summary>
-		[BizImport(CallingConvention.Cdecl)] // the core will handle this itself; you only need to call this when using the cache for your own purposes
-		public abstract void gpgx_flush_vram();
-
-		/// <summary>
-		/// mark the bg pattern cache as dirty
-		/// </summary>
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_invalidate_pattern_cache();
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_flush_vram();
 
 		[StructLayout(LayoutKind.Sequential)]
 		public struct RegisterInfo
@@ -323,11 +323,11 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 			public IntPtr Name;
 		}
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract int gpgx_getmaxnumregs();
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern int gpgx_getmaxnumregs();
 
-		[BizImport(CallingConvention.Cdecl, Compatibility = true)]
-		public abstract int gpgx_getregs([Out] RegisterInfo[] regs);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern int gpgx_getregs([Out] RegisterInfo[] regs);
 
 		[Flags]
 		public enum DrawMask : int
@@ -339,16 +339,17 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx
 			Backdrop = 16
 		}
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_set_draw_mask(DrawMask mask);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_set_draw_mask(DrawMask mask);
 
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_write_m68k_bus(uint addr, byte data);
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract void gpgx_write_s68k_bus(uint addr, byte data);
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract byte gpgx_peek_m68k_bus(uint addr);
-		[BizImport(CallingConvention.Cdecl)]
-		public abstract byte gpgx_peek_s68k_bus(uint addr);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_write_m68k_bus(uint addr, byte data);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern void gpgx_write_s68k_bus(uint addr, byte data);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern byte gpgx_peek_m68k_bus(uint addr);
+		[DllImport("libgenplusgx.dll", CallingConvention = CallingConvention.Cdecl)]
+		public static extern byte gpgx_peek_s68k_bus(uint addr);
+
 	}
 }
