+++ BizHawk.Emulation.Cores/Consoles/Sega/gpgx64/GPGX.IDebuggable.cs
@@ -4,15 +4,15 @@ using System.Runtime.InteropServices;
 
 using BizHawk.Emulation.Common;
 
-namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx32
+namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx64
 {
 	public partial class GPGX : IDebuggable
 	{
 		public IDictionary<string, RegisterValue> GetCpuFlagsAndRegisters()
 		{
-			LibGPGX.RegisterInfo[] regs = new LibGPGX.RegisterInfo[LibGPGX.gpgx_getmaxnumregs()];
+			LibGPGX.RegisterInfo[] regs = new LibGPGX.RegisterInfo[Core.gpgx_getmaxnumregs()];
 
-			int n = LibGPGX.gpgx_getregs(regs);
+			int n = Core.gpgx_getregs(regs);
 			if (n > regs.Length)
 				throw new InvalidOperationException("A buffer overrun has occured!");
 			var ret = new Dictionary<string, RegisterValue>();
@@ -63,7 +63,7 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx32
 
 		private void RefreshMemCallbacks()
 		{
-			LibGPGX.gpgx_set_mem_callback(
+			Core.gpgx_set_mem_callback(
 				MemoryCallbacks.HasReads ? ReadCallback : null,
 				MemoryCallbacks.HasWrites ? WriteCallback : null,
 				MemoryCallbacks.HasExecutes ? ExecCallback : null);
@@ -71,7 +71,7 @@ namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx32
 
 		private void KillMemCallbacks()
 		{
-			LibGPGX.gpgx_set_mem_callback(null, null, null);
+			Core.gpgx_set_mem_callback(null, null, null);
 		}
 	}
 }
