+++ BizHawk.Emulation.Cores/Consoles/Sega/gpgx64/GPGX.ITraceable.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 using BizHawk.Common.NumberExtensions;
 
-namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx32
+namespace BizHawk.Emulation.Cores.Consoles.Sega.gpgx64
 {
 	public partial class GPGX
 	{
